cleanup(ionic-react): break application schematic into separate files
The application schematic was getting rather large, so breaking it down into smaller files was much easier to reason about, and could make it easier to unit test these functions in the future.
